+++ src/bindings/node/node_stub.js
@@ -10,3 +10,5 @@ joo_global_object.fs = {
 joo_global_object.child_process = require("child_process");
 
 joo_global_object.path = require("path");
+
+joo_global_object.os = require("os");